dlna: Better fix for ILP32.
Coding in the evening is often not very effective...

+++ src/dlna/dbus_upnp_list_filler.cc
@@ -178,8 +178,8 @@ ssize_t UPnP::DBusUPnPFiller::fill(ItemProvider<UPnP::ItemData> &item_provider,
         {
             msg_error(ERANGE, LOG_NOTICE,
                       "Got too many child elements from UPnP server "
-                      "(requested %zu, got %lu), ignoring excess elements",
-                      count, static_cast<unsigned long>(num_of_children));
+                      "(requested %zu, got %zu), ignoring excess elements",
+                      count, num_of_children);
             num_of_children = count;
         }
